add loading page to cohorts page
