+++ src/pages/cohort/teachers/index.js
@@ -1,41 +1,9 @@
 import Card from "../../../components/card";
 import './style.css';
 import Teacher from "./teacher";
-import { get, getUserById } from "../../../service/apiClient";
-import { useEffect, useState } from "react";
-import jwtDecode from "jwt-decode";
 
-const Teachers = () => {
 
-    const [teachers, setTeachers] = useState([]);
-
-    useEffect(() => {
-        async function fetchData() {
-            try {
-                const token = localStorage.getItem('token');
-                const { userId } = jwtDecode(token);
-                const user = await getUserById(userId);
-                const data = await get(`cohorts/${user.cohort.id}`);
-
-                const teachers = data.data.cohort.profiles.filter((userid) => userid?.role?.name === "ROLE_TEACHER");
-                setTeachers(teachers || []);
-            } catch (error) {
-                console.error('fetchTeachers() in cohort/teachers/index.js:', error);
-            }
-        }
-
-        fetchData();
-    }, []);
-
-    function getInitials(teacher) {
-        if (!teacher.firstName || !teacher.lastName) return "NA";
-        const firstNameParts = teacher.firstName.trim().split(/\s+/); // split by any number of spaces
-        const lastNameInitial = teacher.lastName.trim().charAt(0);
-        
-        const firstNameInitials = firstNameParts.map(name => name.charAt(0));
-        
-        return (firstNameInitials.join('') + lastNameInitial).toUpperCase();
-    }
+const Teachers = ({ teachers, getInitials }) => {
 
     return (
         <Card>
